python_basics/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
multidict==6.0.4
mysql-connector==2.2.9
SQLAlchemy==2.0.22
typing_extensions==4.8.0
yarl==1.9.2
